+++ apps/fluent-tester/src/FluentTester/TestComponents/CheckboxExperimental/consts.ts
@@ -1,0 +1,2 @@
+export const HOMEPAGE_CHECKBOX_CHECKBOXEXPERIMENTAL = 'Homepage_Checkbox_CheckboxExperimental'; // Button on the homepage that navigates to the selected test page
+export const ExperimentalCheckboxTestPageId = 'CheckboxExperimental_TestPage'; // A component on each specific test page.