+++ packages/components/Tabs/src/TabsItem.settings.ts
@@ -7,7 +7,7 @@ export const tabsItemSelectActionLabel = 'Select a TabsItem';
 export const settings: IComposeSettings<TabsItemType> = [
   {
     tokens: {
-      color: 'neutralForeground3Brand',
+      color: 'neutralStrokeAccessible',
       variant: 'bodyStandard',
       borderWidth: 2,
       borderColor: 'transparent',
@@ -56,8 +56,17 @@ export const settings: IComposeSettings<TabsItemType> = [
       },
       hovered: {
         tokens: {
-          color: 'neutralForeground2Hover',
+          color: 'neutralForeground1',
           indicatorColor: 'neutralStroke1'
+        },
+        _overrides: {
+          selected: {
+            indicator: {
+              style:{
+                marginHorizontal: 0,
+              },
+            },
+          },
         },
       },
       selected: {
@@ -77,7 +86,7 @@ export const settings: IComposeSettings<TabsItemType> = [
 
       pressed: {
         tokens: {
-          color: 'neutralForeground2Pressed',
+          color: 'neutralForeground1',
           indicatorColor: 'brandStroke1',
         },
       },