+++ packages/components/Tabs/src/TabsItem.tsx
@@ -15,7 +15,6 @@ import {
   useAsPressable,
   useViewCommandFocus,
   createIconProps,
-  useOnPressWithFocus,
 } from '@fluentui-react-native/interactive-hooks';
 
 export const TabsItem = compose<TabsItemType>({
@@ -36,30 +35,38 @@ export const TabsItem = compose<TabsItemType>({
       ...rest
     } = userProps;
 
-    // Grabs the context information from Tabs (currently selected TabsItem and client's onTabsClick callback)
+    // Grabs the context information from Tabs (currently selected TabsItem and client's onTabsClick callback).
     const info = React.useContext(TabsContext);
 
-    /* There's a bug where the user callback is being called multiple times on one click.
-    We check that there is an actual change in selection before forwarding the message to the user callback
-    so that they aren't notified more than once for each click. */
-    const changeSelection = () => {
-      if (itemKey != info.selectedKey) {
+    const [focusState, setFocusState] = React.useState({
+      focused: false,
+    });
+
+    const changeSelection = React.useCallback(() => {
+      componentRef?.current?.focus();
+    }, [componentRef]);
+
+    const changeSelectionWithFocus = React.useCallback(() => {
+      setFocusState({ focused: true });
+      if (!focusState.focused) {
         info.onTabsClick && info.onTabsClick(itemKey);
         info.getTabId && info.getTabId(itemKey, info.tabsItemKeys.findIndex(x => x == itemKey) + 1);
         info.updateSelectedTabsItemRef && componentRef && info.updateSelectedTabsItemRef(componentRef);
       }
-    };
+    }, [focusState, setFocusState, componentRef, info, itemKey]);
 
-    // Ensure focus is placed on tabsItem after click
-    const changeSelectionWithFocus = useOnPressWithFocus(componentRef, changeSelection);
+    const removeFocus = React.useCallback(() => {
+      setFocusState({ focused: false });
+    }, [setFocusState]);
 
     const pressable = useAsPressable({
       ...rest,
-      onPress: changeSelectionWithFocus,
-      onFocus: changeSelection,
+      onPress: changeSelection,
+      onFocus: changeSelectionWithFocus,
+      onBlur: removeFocus,
     });
 
-    // set up state
+    // Set up state.
     const state: TabsItemState = {
       info: {
         ...pressable.state,
@@ -84,7 +91,7 @@ export const TabsItem = compose<TabsItemType>({
     // Grab the styling information from the userProps, referencing the state as well as the props.
     const styleProps = useStyling(userProps, (override: string) => state.info[override] || userProps[override]);
 
-    // Used when creating accessibility properties in mergeSettings below
+    // Used when creating accessibility properties in mergeSettings below.
     const onAccessibilityAction = React.useCallback(
       (event: { nativeEvent: { actionName: any } }) => {
         switch (event.nativeEvent.actionName) {
@@ -110,7 +117,7 @@ export const TabsItem = compose<TabsItemType>({
         accessibilityPositionInSet: accessibilityPositionInSet ?? info.tabsItemKeys.findIndex(x => x == itemKey) + 1,
         accessibilitySetSize: accessibilitySetSize ?? info.tabsItemKeys.length,
         onAccessibilityAction: onAccessibilityAction,
-        focusable: Platform.OS === 'macos' ? !userProps.disabled : userProps.focusable ?? true,
+        focusable: Platform.select({default: true, macos: !userProps.disabled}),
       },
       content: { children: headerText + countText, testID: testID },
       icon: createIconProps(icon),
@@ -122,7 +129,7 @@ export const TabsItem = compose<TabsItemType>({
   render: (Slots: ISlots<TabsItemSlotProps>, renderData: TabsItemRenderData, ...children: React.ReactNode[]) => {
     const info = renderData.state!.info;
     const context = React.useContext(TabsContext);
-    // Sets the view that belongs to a TabItem
+    // Sets the view that belongs to a TabItem.
     context.views.set(info.key, children);
 
     return (