+++ packages/components/Tabs/src/TabsItem.types.ts
@@ -27,7 +27,7 @@ export interface TabsItemInfo extends IPressableState {
   selected: boolean;
 
   /**
-   * Key for tabItem
+   * Key for tabItem.
    */
   key: string;
 }
@@ -38,22 +38,22 @@ export interface TabsItemState {
 
 export interface TabsItemProps extends IPressableProps {
   /**
-   * The text string for the option
+   * The text string for the option.
    */
   headerText?: string;
 
   /**
-   * The number for the TabsItem count
+   * The number for the TabsItem count.
    */
   itemCount?: number;
 
   /**
-   * A unique key-identifier for each option
+   * A unique key-identifier for each option.
    */
   itemKey: string;
 
   /**
-   * Whether or not the tabs item is selectable
+   * Whether or not the tabs item is selectable.
    */
   disabled?: boolean;