+++ packages/theming/win32-theme/src/index.ts
@@ -2,3 +2,5 @@ export * from './createOfficeTheme';
 export * from './createPartialOfficeTheme';
 export * from './NativeModule/index';
 export * from './paletteFromOfficeColors';
+export * from './createOfficeAliasTokens';
+export * from './createBrandedThemeWithAlias';