+++ packages/libraries/core/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @fluentui/react-native
 
-This log was last generated on Tue, 31 Aug 2021 18:48:34 GMT and should not be manually modified.
+This log was last generated on Mon, 13 Sep 2021 23:22:43 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.25.14
+
+Mon, 13 Sep 2021 23:22:43 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/menu-button to v0.6.7 (ruaraki@microsoft.com)
+
+## 0.25.13
+
+Fri, 10 Sep 2021 18:16:59 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/menu-button to v0.6.6 (v.kozlova13@gmail.com)
+
+## 0.25.12
+
+Thu, 09 Sep 2021 22:12:19 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/persona to v0.11.15 (ruaraki@microsoft.com)
+
+## 0.25.11
+
+Thu, 09 Sep 2021 20:03:01 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/menu-button to v0.6.4 (v.kozlova13@gmail.com)
 
 ## 0.25.10