+++ packages/theming/theming-utils/CHANGELOG.md
@@ -1,8 +1,44 @@
 # Change Log - @fluentui-react-native/theming-utils
 
-This log was last generated on Thu, 26 Aug 2021 21:02:08 GMT and should not be manually modified.
+This log was last generated on Mon, 13 Sep 2021 23:22:43 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.7.2
+
+Mon, 13 Sep 2021 23:22:43 GMT
+
+### Patches
+
+- Add check for undefined (ruaraki@microsoft.com)
+
+## 0.7.1
+
+Fri, 10 Sep 2021 18:16:59 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-tokens to v0.7.1 (ruaraki@microsoft.com)
+
+## 0.7.0
+
+Thu, 09 Sep 2021 22:12:19 GMT
+
+### Minor changes
+
+- Integrate updated themes (ruaraki@microsoft.com)
+
+### Patches
+
+- Bump @fluentui-react-native/theme-tokens to v0.7.0 (ruaraki@microsoft.com)
+
+## 0.6.1
+
+Thu, 09 Sep 2021 20:03:01 GMT
+
+### Patches
+
+- Reshape input (ruaraki@microsoft.com)
 
 ## 0.6.0